+++ magpie/alembic/versions/2021-06-09_cb92ff1f81bb_add_group_terms.py
@@ -10,8 +10,8 @@ import sqlalchemy as sa
 from alembic import op
 
 # revision identifiers, used by Alembic.
-revision = 'cb92ff1f81bb'
-down_revision = '954a9d7fe740'
+revision = "cb92ff1f81bb"
+down_revision = "954a9d7fe740"
 branch_labels = None
 depends_on = None